Remove old files + make spec works
