--- realworld.cabal
@@ -1,96 +1,0 @@
-name:                realworld
-version:             1.0.0.0
--- synopsis:
--- description:
-homepage:            https://github.com/eckyputrady/haskell-scotty-realworld-example-app#readme
-license:             BSD3
-license-file:        LICENSE
-author:              Ecky Putrady
-maintainer:          eckyputrady@gmail.com
-copyright:           2018 Ecky Putrady
-category:            Web
-build-type:          Simple
-extra-source-files:  README.md
-cabal-version:       >=1.10
-
-library
-  hs-source-dirs:      src
-  exposed-modules:     Lib
-                     , Adapter.HTTP.Client
-                     , Core.Types
-  other-modules:       Core.Services
-                     , Adapter.JWT
-                     , Adapter.PG
-                     , Adapter.HTTP.API
-  build-depends:       base
-                     , classy-prelude
-                     , scotty
-                     , postgresql-simple
-                     , postgresql-simple-migration
-                     , data-has
-                     , aeson
-                     , time
-                     , mtl
-                     , regex-compat
-                     , http-types
-                     , resource-pool
-                     , warp-tls
-                     , warp
-                     , digestive-functors
-                     , digestive-functors-aeson
-                     , lens
-                     , wai
-                     , jose-jwt
-                     , cryptonite
-                     , slug
-                     , convertible
-                     , wai-cors
-                     -- client
-                     , wreq
-                     , http-client
-  default-language:    Haskell2010
-  default-extensions:  NoImplicitPrelude
-                     , OverloadedStrings
-                     , TypeFamilies
-                     , FlexibleContexts
-                     , FlexibleInstances
-                     , ConstraintKinds
-                     , TemplateHaskell
-                     , ScopedTypeVariables
-                     , MultiParamTypeClasses
-                     , QuasiQuotes
-                     , GeneralizedNewtypeDeriving
-  ghc-options:       -Wall -Werror -Wincomplete-record-updates -Wincomplete-uni-patterns -Wredundant-constraints
-
-executable realworld-exe
-  hs-source-dirs:      app
-  main-is:             Main.hs
-  ghc-options:         -threaded -rtsopts -with-rtsopts=-N
-  build-depends:       classy-prelude
-                     , realworld
-  default-language:    Haskell2010
-  default-extensions:  NoImplicitPrelude
-                     , OverloadedStrings
-
-test-suite realworld-test
-  type:                exitcode-stdio-1.0
-  hs-source-dirs:      test
-  main-is:             Main.hs
-  other-modules:       Spec.Common
-                     , Spec.User
-                     , Spec.Article
-                     , Spec.Comment
-  build-depends:       base
-                     , classy-prelude
-                     , realworld
-                     , hspec
-                     , postgresql-simple
-                     , string-random
-  ghc-options:         -threaded -rtsopts -with-rtsopts=-N
-  default-language:    Haskell2010
-  default-extensions:  NoImplicitPrelude
-                     , OverloadedStrings
-
-source-repository head
-  type:     git
-  location: https://github.com/eckyputrady/haskell-scotty-realworld-example-app
