+++ test/Main.hs
@@ -2,7 +2,7 @@ import ClassyPrelude
 import Test.Hspec
 import System.Environment
 import Database.PostgreSQL.Simple
-import qualified Adapter.HTTP.Client as RW
+import qualified Misc.Client as RW
 import qualified Lib
 import Control.Concurrent
