+++ test/Spec/User.hs
@@ -2,8 +2,12 @@ module Spec.User (spec) where
 
 import ClassyPrelude
 import Test.Hspec
-import Core.Types
-import qualified Adapter.HTTP.Client as RW
+import Feature.Comment.Types
+import Feature.User.Types
+import Feature.Article.Types
+import Feature.Auth.Types
+import Feature.Common.Types
+import qualified Misc.Client as RW
 import Text.StringRandom
 import Control.Concurrent
